requirements.txt
# Notion API
requests==2.31.0
python-dotenv==1.0.0

# RSS 파싱 (블로그용)
feedparser==6.0.11

# 데이터 처리 및 시각화
pandas==2.1.4
numpy==1.26.3

# Streamlit 앱용
streamlit==1.29.0
matplotlib==3.8.2
